Ask if overwriting.

+++ bin/nrdp.sh
@@ -31,7 +31,18 @@ while [ -n "$1" ]; do
 done
 
 
-[ -z $build ] && build=`curl "http://builds.netflix.com/view/PPD/view/PPD-NRDAPP/view/all/job/PPD-NRDAPP-$version/" \
+[ -z $build ] && build=`curl "http://builds.netflix.com/view/PPD/view/PPD-NRDAPP/view/all/job/PPD-NRDAPP-$version/" 2> /dev/null \
                         | grep "Last successful build (#[0-9]" \
                         | sed -e 's,^.*Last successful build (#\([0-9]\+\).*$,\1,'`
+if [ -f "$output" ]; then
+    echo -n "$output exists, are you sure you want to overwrite? (Y/N): "
+    read answer
+    case "$answer" in
+        yes|y|Y)
+            ;;
+        *)
+            exit 1
+            ;;
+    esac
+fi
 wget "http://builds.netflix.com/view/PPD/view/PPD-NRDAPP/view/all/job/PPD-NRDAPP-$version/lastSuccessfulBuild/artifact/nrdapp_$version-Debug.tar.gz" -O "$output"
